+++ src/pyfx/view/json_lib/__init__.py
@@ -6,15 +6,23 @@ Example
 .. code-block:: python
    :linenos:
 
-    from pyfx.view.json_lib import JSONListBox, JSONListWalker, NodeFactory
+    from pyfx.view.json_lib import JSONListBox, JSONListWalker, NodeFactory, DEFAULT_NODE_IMPLS
 
-    # 1. create top node from the data (only supports dict, list and primitive variable)
-    top_node = NodeFactory.create_node("", data, display_key=False)
+    # 1. create NodeFactory
+    node_factory = NodeFactory(DEFAULT_NODE_IMPLS)
 
-    # 2. create JSONListBox from top node, a :class:`urwid.ListBox` compatible widget
+    # 2. add any custom display implementation
+    node_factory.register({
+        clazz: json_node_impl
+    })
+
+    # 3. create top node from the data (only supports dict, list and primitive variable)
+    top_node = node_factory.create_root_node(data)
+
+    # 4. create JSONListBox from top node
     listbox = JSONListBox(JSONListWalker(top_node))
 
-    # 3. use listbox in your own TUI
+    # 5. use listbox in your own TUI
 
 Implementation Details
 ======================
@@ -52,3 +60,16 @@ The JSON data is loaded into memory as a tree and based on the data type it crea
 from .json_listbox import JSONListBox
 from .json_listwalker import JSONListWalker
 from .node_factory import NodeFactory
+from .object import ObjectNode
+from .array import ArrayNode
+from .primitive import StringNode, IntegerNode, NumericNode, BooleanNode, NullNode
+
+DEFAULT_NODE_IMPLS = {
+    list: ArrayNode,
+    dict: ObjectNode,
+    str: StringNode,
+    int: IntegerNode,
+    bool: BooleanNode,
+    float: NumericNode,
+    type(None): NullNode
+}