+++ src/pyfx/view/json_lib/object/__init__.py
@@ -1,1 +1,2 @@
 """Node implementation for JSON `object` type."""
+from .object_node import ObjectNode