+++ src/pyfx/view/json_lib/object/object_node.py
@@ -2,7 +2,6 @@ from typing import Union
 
 from overrides import overrides
 
-from .. import node_factory
 from ..json_composite_end_node import JSONCompositeEndNode
 from ..json_composite_node import JSONCompositeNode
 from .object_end_node import ObjectEndNode
@@ -21,13 +20,8 @@ class ObjectNode(JSONCompositeNode):
     * sorted_children_key_list_size: size of key
     """
 
-    def __init__(self,
-                 key: str,
-                 value: dict,
-                 parent: Union["ObjectNode", "array_node", None] = None,
-                 display_key: bool = True
-                 ):
-        super().__init__(key, value, parent, display_key)
+    def __init__(self, key: str, value: dict, node_factory, parent=None, display_key=True):
+        super().__init__(key, value, node_factory, parent, display_key)
         self._children = {}
         self._sorted_children_key_list = sorted(value.keys())
         # avoid re-calculation
@@ -41,17 +35,17 @@ class ObjectNode(JSONCompositeNode):
         self.toggle_expanded()
 
     @overrides
-    def has_children(self) -> bool:
+    def has_children(self):
         return self._sorted_children_key_list_size != 0
 
     @overrides
-    def get_first_child(self) -> Union["JSONSimpleNode", None]:
+    def get_first_child(self):
         if not self.has_children():
             return None
         return self._get_child_node(self._sorted_children_key_list[0])
 
     @overrides
-    def get_last_child(self) -> Union["JSONSimpleNode", None]:
+    def get_last_child(self):
         if not self.has_children():
             return None
         return self._get_child_node(self._sorted_children_key_list[self._sorted_children_key_list_size - 1])
@@ -79,7 +73,7 @@ class ObjectNode(JSONCompositeNode):
 
     def _load_child_node(self, key):
         value = self.get_value()[key]
-        return node_factory.NodeFactory.create_node(key, value, parent=self, display_key=True)
+        return self._node_factory.create_node(key, value, parent=self, display_key=True)
 
     # =================================================================================== #
     # ui                                                                                  #